Change transformers import based on code review
