+++ lib/adapters/default.js
@@ -1,11 +1,7 @@
 const url = require('url');
 const uuidv4 = require('uuid/v4');
 
-const mapOutRes = require('../transformers/mapOutRes');
-const mapInReq = require('../transformers/mapInReq');
-const mapOutReq = require('../transformers/mapOutReq');
-const mapInRes = require('../transformers/mapInRes');
-const mapError = require('../transformers/mapError');
+const { mapError, mapInReq, mapInRes, mapOutReq, mapOutRes } = require('../transformers/transformers.js');
 
 const context = (ctx, traceHeaderName) => {
   return {
