--- lib/transformers/mapError.js
@@ -1,12 +1,0 @@
-const logTags = require('../logTags');
-
-module.exports = ({ ctx, err }) => {
-  err.action = `${ctx.logCtx.method}${ctx.request.path}`;
-  err.params = {
-    query: ctx.request.query,
-    body: ctx.request.body,
-    log_tag: logTags.CATEGORY.UNEXPECTED_ERROR.TAG
-  };
-  err.context = Object.assign(err.context || {}, ctx.logCtx);
-  return err;
-};
